+++ physicsnemo/nn/functional/geometry/mesh_poisson_disk_sample/__init__.py
@@ -15,14 +15,5 @@
 # limitations under the License.
 
 from .mesh_poisson_disk_sample import MeshPoissonDiskSample, mesh_poisson_disk_sample
-from .mesh_to_voxel_fraction import MeshToVoxelFraction, mesh_to_voxel_fraction
-from .sdf import SignedDistanceField, signed_distance_field
 
-__all__ = [
-    "MeshPoissonDiskSample",
-    "MeshToVoxelFraction",
-    "SignedDistanceField",
-    "mesh_poisson_disk_sample",
-    "mesh_to_voxel_fraction",
-    "signed_distance_field",
-]
+__all__ = ["MeshPoissonDiskSample", "mesh_poisson_disk_sample"]